This adds the first unittest for process-interaton-samples
Process-iteration-samples, well, process the iteration samples.
Every benchmark [should] have multiple test samples for any
benchmark iterartion.  The samples for the iteration must be
examined to compute the mean and standard deviation, and
do some houskeeping, like tar'ing up sample data, and
letting the caller know if this iteration passed the max-stddev
threshold.

All benchmarks which use the new JSON metric format can use
process-iteration-samples.  For unittests, all we need are:
1: multiple sample directories, each with a result.json
2: a process-iteration-samples-args file so the unittest
   uses the right arguments (often benchmark specific)
That's it.  As each benchmark adopts the JSON format,
a new samples/process-iteration-samples-N directory
should be created, with multiple benchmark iteration
sample subdirs, each with result.json files for that
benchmark.  The current one is for uperf.